requirements.txt
blinker==1.8.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
flask==3.0.3
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.6
MarkupSafe==2.1.5
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
werkzeug==3.0.6
zipp==3.20.2
